first_version/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
distro==1.9.0
fastapi==0.119.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.11.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==1.0.1
langchain-core==1.0.0
langchain-openai==1.0.0
langgraph==1.0.1
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.1
langgraph-sdk==0.2.9
langsmith==0.4.37
lxml==6.0.2
openai==2.6.0
orjson==3.11.3
ormsgpack==1.11.0
packaging==25.0
psycopg2-binary==2.9.11
pydantic==2.12.3
pydantic_core==2.41.4
pyodbc==5.3.0
python-dotenv==1.1.1
PyYAML==6.0.3
referencing==0.37.0
regex==2025.10.22
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.27.1
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlglot==27.27.0
starlette==0.48.0
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
xxhash==3.6.0
zstandard==0.25.0
